# docs/swagger.yaml
basePath: /
definitions:
...
  service.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  service.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      organization_id:
        minimum: 1
        type: integer
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - organization_id
    - password
    type: object
...
  service.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  service.LoginUserResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/service.UserResponse'
    type: object
...
  service.UpdateUserProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
...
  service.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      organization_id:
        type: integer
      role:
        type: string
      workspace_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goslack.dev
    name: API Support
    url: https://github.com/heyrmi/goslack/issues
  description: A Slack-like collaboration platform API with real-time messaging, file
    sharing, and workspace management.
  license:
    name: MIT
    url: https://github.com/heyrmi/goslack/blob/main/LICENSE
  termsOfService: https://github.com/heyrmi/goslack
  title: GoSlack API
  version: "1.0"
paths:
...
  /users:
    get:
      description: List users in the authenticated user's organization
      parameters:
      - description: 'Page ID (default: 1)'
        in: query
        minimum: 1
        name: page_id
        type: integer
      - description: 'Page size (default: 10, max: 10)'
        in: query
        maximum: 10
        minimum: 5
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/service.UserResponse'
            type: array
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Register a new user in an organization
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Email already exists or invalid organization
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create User
      tags:
      - users
  /users/{id}:
    get:
      description: Retrieve user information by ID (requires authentication)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - different organization
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get User
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Change user password (users can only change their own password)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Password change details
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/service.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or wrong old password
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Can only change own password
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - users
  /users/{id}/profile:
    put:
      consumes:
      - application/json
      description: Update user profile information (users can only update their own
        profile)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile update details
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Can only update own profile
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - users
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and receive access token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/service.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with access token
          schema:
            $ref: '#/definitions/service.LoginUserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"